+++ src/harbor/models/environment_type.py
@@ -17,3 +17,4 @@ class EnvironmentType(str, Enum):
     CWSANDBOX = "cwsandbox"
     WANDB = "wandb"
     USE_COMPUTER = "use-computer"
+    BLAXEL = "blaxel"